Split new Parquet reader modules
